added searchenginebase class
