+++ appliapps/tpp/searchengines/omssa.py
@@ -45,24 +45,6 @@ class Omssa(WrappedApp):
        Default = `2'
     """
 
-    def add_args(self):
-        return [
-            Argument(Keys.EXECUTABLE, KeyHelp.EXECUTABLE),
-            Argument(Keys.WORKDIR, KeyHelp.WORKDIR),
-            Argument(Keys.THREADS, KeyHelp.THREADS),
-            Argument(Keys.MZXML, KeyHelp.MZXML),
-
-            Argument('FRAGMASSERR', 'Fragment mass error'),
-            Argument('FRAGMASSUNIT', 'Unit of the fragment mass error'),
-            Argument('PRECMASSERR', 'Precursor mass error'),
-            Argument('PRECMASSUNIT', 'Unit of the precursor mass error'),
-            Argument('MISSEDCLEAVAGE', 'Number of maximal allowed missed cleavages'),
-            Argument('DBASE', 'Sequence database file with target/decoy entries'),
-            Argument('ENZYME', 'Enzyme used to digest the proteins'),
-            Argument('STATIC_MODS', 'List of static modifications'),
-            Argument('VARIABLE_MODS', 'List of variable modifications'),
-        ]
-
     def prepare_run(self, log, info):
         wd = info[Keys.WORKDIR]
